post review

+++ src/main/java/org/gridsuite/modification/server/modifications/ByFormulaModification.java
@@ -266,7 +266,7 @@ public class ByFormulaModification extends AbstractModification {
             try {
                 final Double newValue = applyOperation(formulaInfos.getOperator(), value1, value2);
                 switch (identifiable.getType()) {
-                    case GENERATOR -> GeneratorField.setNewValue((Generator) identifiable, formulaInfos.getEditedField(), newValue);
+                    case GENERATOR -> GeneratorField.setNewValue((Generator) identifiable, formulaInfos.getEditedField(), newValue.doubleValue());
                     case BATTERY -> BatteryField.setNewValue((Battery) identifiable, formulaInfos.getEditedField(), newValue);
                     case SHUNT_COMPENSATOR -> ShuntCompensatorField.setNewValue((ShuntCompensator) identifiable, formulaInfos.getEditedField(), newValue);
                     case VOLTAGE_LEVEL -> VoltageLevelField.setNewValue((VoltageLevel) identifiable, formulaInfos.getEditedField(), newValue);
